test: cover project generation module
