+++ src/projectGeneration/index.ts
@@ -1,0 +1,5 @@
+export { registerProjectGenerationCommands } from "./commands";
+export { scaffoldProjectRequest } from "./scaffold";
+export { handleProjectScaffoldUri, setProjectScaffoldListener } from "./uri";
+export { applyTemplate } from "./applyTemplate";
+export * from "./templates";
